+++ lib/tooling/base-tool.js
@@ -26,6 +26,7 @@
 const
     exec = require('../exec'),
     utils = require('../utils'),
+    _ = require('underscore'),
     logger = require('../logger').logger,
     path = require('path');
 
@@ -68,6 +69,27 @@ class BaseTool {
             languageId: "stderr",
             stdout: utils.parseOutput(message)
         };
+    }
+
+    // mostly copy&paste from base-compiler.js
+    findLibVersion(selectedLib, compiler) {
+        const foundLib = _.find(compiler.libs, (o, libId) => libId === selectedLib.id);
+        if (!foundLib) return false;
+
+        const foundVersion = _.find(foundLib.versions, (o, versionId) => versionId === selectedLib.version);
+        return foundVersion;
+    }
+
+    // mostly copy&paste from base-compiler.js
+    getIncludeArguments(libraries, compiler) {
+        const includeFlag = "-I";
+
+        return _.flatten(_.map(libraries, (selectedLib) => {
+            const foundVersion = this.findLibVersion(selectedLib, compiler);
+            if (!foundVersion) return false;
+
+            return _.map(foundVersion.path, (path) => includeFlag + path);
+        }));
     }
 
     runTool(compilationInfo, inputFilepath, args, stdin) {
