+++ lib/tooling/clang-tidy-tool.js
@@ -26,7 +26,6 @@
 const
     fs = require('fs-extra'),
     path = require('path'),
-    _ = require('underscore'),
     BaseTool = require('./base-tool');
 
 class ClangTidyTool extends BaseTool {
@@ -34,32 +33,11 @@ class ClangTidyTool extends BaseTool {
         super(toolInfo, env);
     }
 
-    // mostly copy&paste from bse-compiler.js
-    findLibVersion(selectedLib, compiler) {
-        const foundLib = _.find(compiler.libs, (o, libId) => libId === selectedLib.id);
-        if (!foundLib) return false;
-
-        const foundVersion = _.find(foundLib.versions, (o, versionId) => versionId === selectedLib.version);
-        return foundVersion;
-    }
-
-    // mostly copy&paste from bse-compiler.js
-    getIncludeArguments(libraries, compiler) {
-        const includeFlag = "-I";
-
-        return _.flatten(_.map(libraries, (selectedLib) => {
-            const foundVersion = this.findLibVersion(selectedLib, compiler);
-            if (!foundVersion) return false;
-
-            return _.map(foundVersion.path, (path) => includeFlag + path);
-        }));
-    }
-
     runTool(compilationInfo, inputFilepath, args) {
         const sourcefile = inputFilepath;
         const options = compilationInfo.options;
         const dir = path.dirname(sourcefile);
-        const includeflags = this.getIncludeArguments(compilationInfo.libraries, compilationInfo.compiler);
+        const includeflags = super.getIncludeArguments(compilationInfo.libraries, compilationInfo.compiler);
 
         let source;
         const wantsFix = args.find(option => option.includes("-fix"));
